Implement POSStoresProviding
It only requires a single dispatch method that is used within POS application
